add test case for httpoberver
