+++ client/lookout-reg-server/src/main/java/com/alipay/lookout/remote/step/LookoutRegistry.java
@@ -22,6 +22,7 @@ import com.alipay.lookout.core.CommonTagsAccessor;
 import com.alipay.lookout.core.config.LookoutConfig;
 import com.alipay.lookout.remote.model.LookoutMeasurement;
 import com.alipay.lookout.remote.report.*;
+import com.alipay.lookout.remote.report.support.ReportDecider;
 import com.alipay.lookout.report.MetricObserver;
 
 import java.util.Collection;
@@ -48,7 +49,8 @@ public final class LookoutRegistry extends StepRegistry implements CommonTagsAcc
                            LookoutConfig config) {
         super(clock, config);
         if (observer == null) {
-            observer = new HttpObserver(config, getAddressService(config), this);
+            observer = new HttpObserver(config, getAddressService(config), this,
+                new ReportDecider());
         }
         addMetricObserver(observer);
         this.poller = new SchedulerPoller(this, config, metricObserverComposite);
