+++ client/lookout-reg-server/src/test/java/com/alipay/lookout/remote/report/support/HttpRequestProcessorTest.java
@@ -16,7 +16,7 @@
  */
 package com.alipay.lookout.remote.report.support;
 
-import com.alipay.lookout.remote.report.support.http.HttpRequestProcessor;
+import com.alipay.lookout.remote.report.support.http.DefaultHttpRequestProcessor;
 import org.apache.http.Header;
 import org.apache.http.HttpResponse;
 import org.apache.http.StatusLine;
@@ -24,14 +24,14 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Mockito;
 
-import static com.alipay.lookout.remote.report.support.http.HttpRequestProcessor.WAIT_MINUTES;
+import static com.alipay.lookout.remote.report.support.http.DefaultHttpRequestProcessor.WAIT_MINUTES;
 
 /**
  * Created by kevin.luy@alipay.com on 2017/4/13.
  */
 public class HttpRequestProcessorTest {
     final ReportDecider        reportDecider        = new ReportDecider();
-    final HttpRequestProcessor httpRequestProcessor = new HttpRequestProcessor(reportDecider);
+    final DefaultHttpRequestProcessor httpRequestProcessor = new DefaultHttpRequestProcessor(reportDecider);
 
     @Test
     public void testHandleErrorResponse401() {
